Refactor Confluence models into modular directory structure
