+++ src/mcp_atlassian/models/__init__.py
@@ -9,8 +9,9 @@ simplified dictionaries for API responses.
 # Re-export models for easier imports
 from .base import ApiModel, TimestampMixin
 
-# Confluence models
+# Confluence models (Import from the new structure)
 from .confluence import (
+    ConfluenceAttachment,
     ConfluenceComment,
     ConfluencePage,
     ConfluenceSearchResult,
@@ -18,7 +19,7 @@ from .confluence import (
     ConfluenceUser,
     ConfluenceVersion,
 )
-from .constants import (
+from .constants import (  # noqa: F401 - Keep constants available
     CONFLUENCE_DEFAULT_ID,
     CONFLUENCE_DEFAULT_SPACE,
     CONFLUENCE_DEFAULT_VERSION,
@@ -35,7 +36,7 @@ from .constants import (
     UNKNOWN,
 )
 
-# Jira models
+# Jira models (Keep existing imports)
 from .jira import (
     JiraAttachment,
     JiraBoard,
@@ -100,4 +101,5 @@ __all__ = [
     "ConfluenceComment",
     "ConfluenceVersion",
     "ConfluenceSearchResult",
+    "ConfluenceAttachment",
 ]
